requirements.txt
requests
python-dotenv
rich
click
gunicorn